Add delete key test
