+++ test/functional/ios/testapp/keyboard-specs.js
@@ -2,9 +2,10 @@
 
 var setup = require("../../common/setup-base")
  ,  desired = require('./desired')
- ,  _ = require('underscore');
+ ,  _ = require('underscore')
+ , unorm = require('unorm');
 
-describe('testapp', function () {
+describe('testapp - keyboard', function () {
   var ctx = this;
 
   var test = function (strategy) {
@@ -37,35 +38,65 @@ describe('testapp', function () {
     });
   };
 
-  test();
-  test('oneByOne');
-  test('grouped');
-  test('setValue');
+  _.each([undefined, 'oneByOne', 'grouped', 'setValue'], test);
 
-  describe("keyboard stability @skip-ci", function () {
-    var runs = 10
-      , text = 'Delhi is New @@@ BREAKFAST-FOOD-0001';
-
+  describe("typing", function () {
     var driver;
     setup(this, _.defaults({
       deviceName: 'iPad 2',
     }, desired)).then(function (d) { driver = d; });
 
-    var test = function () {
-      it("should send keys to a text field", function (done) {
-        driver
-          .elementByClassName('UIATextField')
-            .clear()
-            .sendKeys(text)
-            .text().should.become(text)
-          .nodeify(done);
-      });
-    };
+    describe("stability @skip-ci", function () {
+      var runs = 10
+        , text = 'Delhi is New @@@ BREAKFAST-FOOD-0001';
 
-    for (var n = 0; n < runs; n++) {
-      describe('sendKeys test ' + (n + 1), test);
-    }
+      var test = function () {
+        it("should send keys to a text field", function (done) {
+          driver
+            .elementByClassName('UIATextField')
+              .clear()
+              .sendKeys(text)
+              .text().should.become(text)
+            .nodeify(done);
+        });
+      };
+
+      for (var n = 0; n < runs; n++) {
+        describe('sendKeys test ' + (n + 1), test);
+      }
+    });
+
+
+    it('should send accented text', function (done) {
+      var testText = unorm.nfd("é Œ ù ḍ");
+      driver
+        .elementsByClassName('UIATextField').at(1)
+          .sendKeys(testText)
+          .text()
+          .should.become(testText)
+        .nodeify(done);
+    });
+
+    it('should send backspace key', function (done) {
+      driver
+        .elementsByClassName('UIATextField').at(1)
+          .clear()
+          .sendKeys("abcd")
+          .sendKeys('\uE003\uE003')
+          .text()
+          .should.become("ab")
+        .nodeify(done);
+    });
+
+    it('should send delete key', function (done) {
+      driver
+        .elementsByClassName('UIATextField').at(1)
+          .clear()
+          .sendKeys("abcd")
+          .sendKeys('\ue017\ue017')
+          .text()
+          .should.become("ab")
+        .nodeify(done);
+    });
   });
-
-
 });
